solve some issues in the create and edit cabins

+++ src/services/apiCabins.js
@@ -1,4 +1,4 @@
-import supabase from "./supabase";
+import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
@@ -8,30 +8,39 @@ export async function getCabins() {
   }
   return data;
 }
-export async function createCabin(newCabin) {
+
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+  // typeof newCabin.image === "string" &&
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  let query = supabase.from("cabins");
 
-  const imagePath = `${
-    import.meta.env.VITE_SUPABASE_URL
-  }/storage/v1/object/public/cabin-images/${imageName}`;
+  // create / edit cabin
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be created");
   }
+
   //2 upload image
+  if (hasImagePath) return data; ////////----1
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
   //3 delete the cabin  if  there was an error in uploading image
-   if (storageError) {
+  if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     if (error) {
       console.error(storageError);
